pandas-tab1/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
gspread>=5.0.0
google-auth-oauthlib>=0.7.0
google-auth-httplib2>=0.1.0
google-auth>=2.0.0
